LPS-54323 SF - move to web folder
