+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/lar/PasswordPolicyPortletDataHandler.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.password.policies.admin.lar;
+package com.liferay.password.policies.admin.web.lar;
 
-import com.liferay.password.policies.admin.constants.PasswordPoliciesAdminPortletKeys;
+import com.liferay.password.policies.admin.web.constants.PasswordPoliciesAdminPortletKeys;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.lar.BasePortletDataHandler;
