+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/portlet/PasswordPoliciesAdminPortlet.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.password.policies.admin.portlet;
+package com.liferay.password.policies.admin.web.portlet;
 
-import com.liferay.password.policies.admin.constants.PasswordPoliciesAdminPortletKeys;
-import com.liferay.password.policies.admin.upgrade.PasswordPoliciesAdminUpgrade;
+import com.liferay.password.policies.admin.web.constants.PasswordPoliciesAdminPortletKeys;
+import com.liferay.password.policies.admin.web.upgrade.PasswordPoliciesAdminUpgrade;
 import com.liferay.portal.DuplicatePasswordPolicyException;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 import com.liferay.portal.kernel.servlet.SessionErrors;
