+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/search/OrganizationPasswordPolicyChecker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.password.policies.admin.search;
+package com.liferay.password.policies.admin.web.search;
 
 import com.liferay.portal.kernel.dao.search.RowChecker;
 import com.liferay.portal.kernel.log.Log;
