+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/util/PasswordPoliciesAdminImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.password.policies.admin.util;
+package com.liferay.password.policies.admin.web.util;
 
 import com.liferay.portal.kernel.security.pacl.DoPrivileged;
 import com.liferay.portal.kernel.util.OrderByComparator;
